new class create tables
